🏷️ [ADD] 일기 작성, 조회 연동 코드 추가

+++ facemind/lib/api/model/home_calendar_response.dart
@@ -24,7 +24,7 @@ class HomeCalendarData {
 
 class CalendarResult {
   final String date;
-  final int stressRate;
+  final double stressRate;
 
   CalendarResult({required this.date, required this.stressRate});
